+++ src/main/java/com/yisiliang/idea/plugins/apusic/utils/PluginConstant.java
@@ -1,0 +1,17 @@
+package com.yisiliang.idea.plugins.apusic.utils;
+
+public class PluginConstant {
+
+    public static final String DOMAIN_KEY = "domain";
+    public static final String APUSIC_SERVER_KEY = "apusic-server";
+    public static final String DEPLOYMENT_DIRECTORY_KEY = "deployment-directory";
+    public static final String CONTEXT_PATH_KEY = "context-path";
+    public static final String URL_KEY = "url";
+    public static final String ADD_LIBRARIES_AND_CLASSES_KEY = "add-libraries-and-classes";
+    public static final String EXTERNAL_CLASSPATH_KEY = "external-classpath";
+    public static final String VM_OPTIONS_KEY = "vm-options";
+    public static final String ENV_OPTIONS_KEY = "env-options";
+    public static final String DOC_BASE_BROWSE_TITLE_KEY = "doc-base-browse-title";
+    public static final String DOC_BASE_BROWSE_DESC_KEY = "doc-base-browse-desc";
+    public static final String ADD_CHECKBOX_TOOL_TIP_TEXT_KEY = "add-libs-classes-tool-tip";
+}