+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
 # SmartApusic Changelog
+## [0.0.8] - 2023-02-08
+### Added
+
+- Chinese language support.
+- config.xml changed when necessary.
 
 ## [0.0.7] - 2023-02-01
 ### Fixed